requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
black==24.3.0
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-ninja==1.1.0
injector==0.21.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
